consolidate utility files
