--- src/util.ts
@@ -1,4 +1,0 @@
-export function parseLatLon(latLngStr: string): [number, number] {
-  const ll = latLngStr.split(',');
-  return [parseFloat(ll[0]), parseFloat(ll[1])];
-}
